Added Powerups to go with effects: - powerups have an effect which is automatically added to EffectManager in OnCollect - the above can be overridden for custom effects (see PowerupMagnetism.cpp)
Added implementations of powerups for magnetism, pull and smash.

+++ Sustainable Economy/Sustainable Economy/Throwable.h
@@ -55,7 +55,7 @@ private:
 
 	// Effects
 	int m_homingDistance; // The distance at which the throwable will stop homing
-	int m_homingSpeed; // The pixels by which the throwable homes in each frame
+	float m_homingSpeed; // The pixels by which the throwable homes in each frame
 	bool m_isBouncy;
 
 	//Kinematics
